test_types_database.py now passes all tests.
Added the new required missing schema_migrations_module to several
typedefs, and also fixed the database model which was assuming that
there would be at least one migration. I wrote a task in Rally to follow
up and make sure we have a unit test that verifies that we handle the
zero migrations case correctly.